More stable against all 3 browsers

+++ src/test/java/integrationTests/selenium/Selenium_core.java
@@ -46,13 +46,13 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Select;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-
 //import com.smartbear.ready.cmd.runner.SoapUITestCaseRunner;
 
 public class Selenium_core{
 
 	public static WebDriver webdriver;
 	protected static WebDriverWait wait;
+	protected static WebDriverWait quickWait;
 
 	public static String os; 
 	public static String env; 
@@ -134,17 +134,17 @@ public class Selenium_core{
 
 			System.setProperty(FirefoxDriver.SystemProperty.DRIVER_USE_MARIONETTE,"true");
 			System.setProperty(FirefoxDriver.SystemProperty.BROWSER_LOGFILE,"/dev/null");
-			
+
 			if (browserHeadless){
 
-			    FirefoxBinary firefoxBinary = new FirefoxBinary();
-			    firefoxBinary.addCommandLineOptions("--headless");
-			    FirefoxOptions firefoxOptions = new FirefoxOptions();
-			    firefoxOptions.setBinary(firefoxBinary);
-			    webdriver = new FirefoxDriver(firefoxOptions);	
-			    webdriver.manage().window().setSize(new Dimension(1080, 1920));
-			    
-	
+				FirefoxBinary firefoxBinary = new FirefoxBinary();
+				firefoxBinary.addCommandLineOptions("--headless");
+				FirefoxOptions firefoxOptions = new FirefoxOptions();
+				firefoxOptions.setBinary(firefoxBinary);
+				webdriver = new FirefoxDriver(firefoxOptions);	
+				webdriver.manage().window().setSize(new Dimension(1080, 1920));
+
+
 			}else{
 
 				webdriver = new FirefoxDriver();
@@ -155,39 +155,42 @@ public class Selenium_core{
 		case "ie":
 		case "edge":
 
-			
+
 			//Cucumber_runConfig_IT a = new Cucumber_runConfig_IT();
-			
+
 			if (browserHeadless){
 
 				System.out.println("===========================");
 				System.out.println(browser + " doesn't have a headless mode, launching normally");
 				System.out.println("===========================");
-				
+
 			}	
-				
+
 			webdriver = new  EdgeDriver();
+			webdriver.manage().window().maximize();
 
 			break;
 
 
 		default: 
-	
+
 			System.out.println("===========================");
 			System.out.println(browser + " is not a recognised web browser, please check config. Aborting test");
 			System.out.println("===========================");
 			System.exit(0);
-	
+
 		}
 
+		wait = new WebDriverWait(webdriver,15);
+		quickWait = new WebDriverWait(webdriver, 1);
 
-		wait = new WebDriverWait(webdriver,60);
 		deleteCookies();
+
 
 	}
 
 	public static void quitWebDriver() throws Exception{
- 
+
 		webdriver.quit();
 
 	}
@@ -227,7 +230,7 @@ public class Selenium_core{
 
 			//[Fail-safe] check its clickable, to ensure it really does exist.
 			try{
-				WebDriverWait quickWait = new WebDriverWait(webdriver, 1);
+
 				quickWait.until(ExpectedConditions.elementToBeClickable(target));
 
 				return true;
@@ -244,13 +247,38 @@ public class Selenium_core{
 	public void waitForElement(By target) throws Exception{
 
 		waitForAjaxComplete();
-		wait.until(ExpectedConditions.visibilityOfElementLocated(target));
 
+		try{
+			wait.until(ExpectedConditions.visibilityOfElementLocated(target));
+		}
+		catch (Exception e){
+			System.out.println("Selenium has waiting its max time for the following element to be visible" + target );	
+		}
 	}
 
 	public void waitForElementInvisible(By target) throws Exception{
 
-		wait.until(ExpectedConditions.invisibilityOfElementLocated(target));
+		waitForAjaxComplete();
+		
+		try{
+			wait.until(ExpectedConditions.invisibilityOfElementLocated(target));
+		}catch(Exception e){
+
+			System.out.println("Selenium has waiting its max time for the following element to not be visible" + target );	
+		}
+
+	}
+
+	public void waitForElementNotClickable(By target) throws Exception{
+
+		waitForAjaxComplete();
+		
+		try{
+			wait.until(ExpectedConditions.not(ExpectedConditions.elementToBeClickable(target)));
+		}catch(Exception e){
+
+			System.out.println("Selenium has waiting its max time for the following element to not be clickable: " + target );	
+		}
 
 	}
 
@@ -271,7 +299,7 @@ public class Selenium_core{
 			Thread.sleep(500);
 			iWaitTime++;
 
-			System.out.println(iWaitTime + "/" + iWaitFinish + " Waiting for page to load (AJAX not included)");
+			//System.out.println(iWaitTime + "/" + iWaitFinish + " Waiting for page to load (AJAX not included)");
 
 			//fail-safe 
 			if (iWaitTime==iWaitFinish){break;}
@@ -294,7 +322,7 @@ public class Selenium_core{
 				Thread.sleep(500);
 				iWaitTime++;
 
-				System.out.println(iWaitTime + "/" + iWaitFinish + " Waiting for active AJAX calls to complete");
+				//System.out.println(iWaitTime + "/" + iWaitFinish + " Waiting for active AJAX calls to complete");
 
 				//fail-safe 
 				if (iWaitTime==iWaitFinish){break;}        
@@ -308,29 +336,36 @@ public class Selenium_core{
 	// Wait for DOM ready and Ajax calls on page to complete (End)
 	//==================================================
 
+	public boolean textExists(String text) throws InterruptedException {
 
-	public boolean textExists(String text) {
+		waitForAjaxComplete();
 
 		return webdriver.getPageSource().toLowerCase().contains(text.toLowerCase());
-
 
 	}
 
 	public void click(By target) throws Exception{
 
+		waitForAjaxComplete();
 
-		wait.until(ExpectedConditions.elementToBeClickable(target)).click();
+		try{
+			wait.until(ExpectedConditions.elementToBeClickable(target));
+		}catch(Exception e){
+			System.out.println("Selenium has waiting its max time for the following element to be clickable: " + target );	
+		}finally{
+			webdriver.findElement(target).click();
+		}
 
 	}
 
-	public void sendkeys(By target,String textToSend){
+	public void sendkeys(By target,String textToSend) throws Exception{
 
-		wait.until(ExpectedConditions.visibilityOfElementLocated(target));
+		waitForElement(target);
 
 		try{
 
 			//Clear text field if it has text before sending text.
-			if(!webdriver.findElement(target).getAttribute("value").equals("") ||
+			if(!webdriver.findElement(target).getAttribute("innerHTML").equals("") ||
 					!webdriver.findElement(target).getText().equals("")){
 
 				webdriver.findElement(target).clear();
@@ -339,28 +374,30 @@ public class Selenium_core{
 		}finally{
 			webdriver.findElement(target).sendKeys(textToSend);	
 		}	
-		
-		
-		webdriver.findElement(target).sendKeys(textToSend);	
+
 
 	}
 
-	public String getText(By target){
+	public String getText(By target) throws Exception{
+
+		waitForElement(target);
 
 		return webdriver.findElement(target).getText();
 
 	}	
 
-	public String getInnerHTML(By target){
+	public String getInnerHTML(By target) throws Exception{
+
+		waitForElement(target);
 
 		return webdriver.findElement(target).getAttribute("innerHTML");
 
 	}	
 
 
-	public void selectByIndex(By target,int index){
+	public void selectByIndex(By target,int index) throws Exception{
 
-		wait.until(ExpectedConditions.visibilityOfElementLocated(target));
+		waitForElement(target);
 
 		Select select = new Select(webdriver.findElement(target));
 		select.selectByIndex(index);
@@ -368,9 +405,9 @@ public class Selenium_core{
 	}
 
 
-	public void selectByVisibleText(By target,String text) throws InterruptedException{
+	public void selectByVisibleText(By target,String text) throws Exception{
 
-		wait.until(ExpectedConditions.visibilityOfElementLocated(target));
+		waitForElement(target);
 
 		Select select = new Select(webdriver.findElement(target));
 		select.selectByVisibleText(text);
@@ -387,9 +424,9 @@ public class Selenium_core{
 
 	}
 
-	public String getDropDownMenuText(By target) {
+	public String getDropDownMenuText(By target) throws Exception {
 
-		wait.until(ExpectedConditions.visibilityOfElementLocated(target));
+		waitForElement(target);
 
 		Select select = new Select(webdriver.findElement(target));
 
@@ -435,7 +472,18 @@ public class Selenium_core{
 
 	public static void deleteCookies(){
 
+		if (webdriver.getCurrentUrl().equals("data:,") || 
+				webdriver.getCurrentUrl().equals("about:blank")){
+
+			return;
+		}
+
 		webdriver.manage().deleteAllCookies();
+
+
+		//[Fail-Safe] make sure cookies cleared
+		//		((JavascriptExecutor) webdriver).executeScript("window.sessionStorage.clear();");
+		//		((JavascriptExecutor) webdriver).executeScript("window.localStorage.clear();");
 
 	}
 
@@ -446,12 +494,9 @@ public class Selenium_core{
 
 	public void scrollTo(By target) throws Exception {
 
-		waitForElement(target);
-
 		WebElement element = webdriver.findElement(target);
 		((JavascriptExecutor) webdriver).executeScript("arguments[0].scrollIntoView(true);", element);
-		//Thread.sleep(1500); 
-
+		Thread.sleep(100); 
 
 	}
 
@@ -462,7 +507,6 @@ public class Selenium_core{
 
 	}
 
-
 	public void scrollBottom() throws InterruptedException {
 
 		((JavascriptExecutor) webdriver).executeScript("window.scrollTo(0, document.body.scrollHeight)");
@@ -471,7 +515,6 @@ public class Selenium_core{
 	}
 
 	public void scrollTop() throws InterruptedException {
-
 
 		((JavascriptExecutor) webdriver).executeScript("window.scrollTo(0, 0)");
 		Thread.sleep(1500);
@@ -482,16 +525,15 @@ public class Selenium_core{
 	// - Scrolling  (End)
 	//================================================
 
-	public void mouseTo(By target) throws InterruptedException {
+	public void mouseTo(By target) throws Exception {
 
-		wait.until(ExpectedConditions.elementToBeClickable(target));
+		waitForElement(target);
 
 		Actions action = new Actions(webdriver);
 		action.moveToElement(webdriver.findElement(target)).build().perform();
 		//Thread.sleep(1500);
 
 	}	
-
 
 	public void highLightElement(By by)  {
 
@@ -556,7 +598,6 @@ public class Selenium_core{
 		System.out.println("==================================================");
 		System.out.println(scriptCounter + " scripts executed by " + webdriver.getCurrentUrl());
 
-
 	}	
 
 
@@ -566,7 +607,6 @@ public class Selenium_core{
 		return  (Boolean) ((JavascriptExecutor)webdriver).executeScript("return arguments[0].complete && typeof arguments[0].naturalWidth != \"undefined\" && arguments[0].naturalWidth > 0", ImageFile);
 
 	}
-
 
 	//=====================================================
 	/*
